ADD component Gretting and set it as a route
