+++ package.json
@@ -16,6 +16,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
+    "react-router-dom": "^6.10.0",
     "redux": "^4.2.1",
     "webpack": "^5.79.0"
   },
